+++ bacnet.c
@@ -186,7 +186,7 @@ PHP_MINFO_FUNCTION(bacnet) {
 	php_info_print_table_start();
 	php_info_print_table_header(2, "BACnet/IP support", "enabled");
 	php_info_print_table_row(2, "Extension version", PHP_BACNET_VERSION);
-	php_info_print_table_row(2, "bacnet-stack", "1.5.0 (5afc5c9a)");
+	php_info_print_table_row(2, "bacnet-stack", "1.5.1 (3a74c74a)");
 	php_info_print_table_end();
 
 	DISPLAY_INI_ENTRIES();